+++ stars-core/src/main/kotlin/tools/aqua/stars/core/types/TickDifference.kt
@@ -39,4 +39,19 @@ abstract class TickDifference<D : TickDifference<D>> : Comparable<D>, Serializab
    * @return A new [TickDifference] object.
    */
   abstract operator fun minus(other: D): D
+
+  /** Holder of extension functions for [TickDifference] objects. */
+  companion object {
+    /** Sums all [TickDifference] objects in the given [Iterable], or returns `null` if empty. */
+    fun <T : TickDifference<T>> Iterable<T>.sumOrNull(): T? =
+        this.reduceOrNull { acc, d -> acc + d }
+
+    /**
+     * Sums all [TickDifference] objects in the given [Iterable].
+     *
+     * @throws IllegalStateException if the [Iterable] is empty.
+     */
+    fun <T : TickDifference<T>> Iterable<T>.sum(): T =
+        checkNotNull(sumOrNull()) { "Cannot sum an empty Iterable of TickDifference." }
+  }
 }